Black and clean up tests.
Reformatted with Black.

Removed unused test files.
