--- tests/response_image_611EovQ.json
@@ -1,26 +1,0 @@
-{
-  "data": {
-    "in_gallery": false,
-    "nsfw": null,
-    "account_url": null,
-    "section": null,
-    "datetime": 1465540781,
-    "width": 4000,
-    "size": 2863896,
-    "bandwidth": 446767776,
-    "is_ad": false,
-    "views": 156,
-    "vote": null,
-    "id": "611EovQ",
-    "link": "http://i.imgur.com/611EovQ.jpg",
-    "type": "image/jpeg",
-    "animated": false,
-    "description": "Right before I moved desks for the 6th time in 1.5 years. I lost my nice window desk, oh well.",
-    "favorite": false,
-    "account_id": null,
-    "height": 2992,
-    "title": "Work, June 1st, 2016: Uber"
-  },
-  "success": true,
-  "status": 200
-}
