+++ tests/test_embed.py
